FIX Adhere to ANSI C

+++ src/dis_pari.c
@@ -4,13 +4,13 @@
 #include <stdio.h>
 
 int main(int argc, char *argv[]) {
-  double x;
+  int x;
 
   printf("Inserisci un intero positivo\n");
-  scanf("%f", &x);
+  scanf("%d", &x);
 
   printf("%f e': ", x);
-  if (x % 2.0 == 0) {
+  if (x % 2 == 0) {
     printf("pari");
   }
   else {
